apply to folder
